spark-batch/requirements.txt
pyspark==3.5.0
psycopg2-binary==2.9.9
boto3==1.34.0
